python/airflow_dags.py
"""
Apex Motors - Apache Airflow DAGs
Orquestração do pipeline de dados ETL
"""

from datetime import datetime, timedelta
from airflow import DAG
from airflow.operators.python import PythonOperator
from airflow.providers.google.cloud.operators.bigquery import BigQueryCreateEmptyTableOperator
from airflow.providers.postgres.operators.postgres import PostgresOperator
from airflow.utils.task_group import TaskGroup

# ============================================================================
# CONFIGURAÇÕES PADRÃO
# ============================================================================

default_args = {
    'owner': 'data-engineering',
    'retries': 2,
    'retry_delay': timedelta(minutes=5),
    'start_date': datetime(2023, 1, 1),
    'email_on_failure': True,
    'email': ['data-team@apexmotors.com'],
}

# ============================================================================
# DAG PRINCIPAL - PIPELINE DIÁRIO
# ============================================================================

dag = DAG(
    'apex_motors_daily_pipeline',
    default_args=default_args,
    description='Pipeline diário de ETL para Apex Motors',
    schedule_interval='0 2 * * *',  # Executa diariamente às 2 AM
    catchup=False,
    tags=['apex-motors', 'production'],
)

# ============================================================================
# FUNÇÕES PYTHON
# ============================================================================

def extract_from_postgres():
    """Extrai dados do PostgreSQL"""
    print("Iniciando extração de dados do PostgreSQL...")
    # Implementar lógica de extração
    print("Extração concluída!")

def validate_data():
    """Valida qualidade dos dados"""
    print("Iniciando validação de dados...")
    # Implementar lógica de validação
    print("Validação concluída!")

def transform_and_load():
    """Transforma dados com dbt e carrega no BigQuery"""
    print("Iniciando transformação com dbt...")
    # Implementar lógica de transformação
    print("Transformação concluída!")

def generate_alerts():
    """Gera alertas baseado em métricas"""
    print("Gerando alertas...")
    # Implementar lógica de alertas
    print("Alertas gerados!")

# ============================================================================
# TASKS
# ============================================================================

# Task 1: Extração
extract_task = PythonOperator(
    task_id='extract_from_postgres',
    python_callable=extract_from_postgres,
    dag=dag,
)

# Task 2: Validação
validate_task = PythonOperator(
    task_id='validate_data',
    python_callable=validate_data,
    dag=dag,
)

# Task 3: Transformação
transform_task = PythonOperator(
    task_id='transform_and_load',
    python_callable=transform_and_load,
    dag=dag,
)

# Task 4: Alertas
alerts_task = PythonOperator(
    task_id='generate_alerts',
    python_callable=generate_alerts,
    dag=dag,
)

# ============================================================================
# DEFINIÇÃO DE DEPENDÊNCIAS
# ============================================================================

extract_task >> validate_task >> transform_task >> alerts_task

# ============================================================================
# DAG SECUNDÁRIA - LIMPEZA DE DADOS
# ============================================================================

dag_cleanup = DAG(
    'apex_motors_data_cleanup',
    default_args=default_args,
    description='Limpeza e manutenção de dados',
    schedule_interval='0 4 * * 0',  # Executa todo domingo às 4 AM
    catchup=False,
    tags=['apex-motors', 'maintenance'],
)

cleanup_task = PythonOperator(
    task_id='cleanup_old_data',
    python_callable=lambda: print("Limpando dados antigos..."),
    dag=dag_cleanup,
)

# ============================================================================
# DAG SECUNDÁRIA - RELATÓRIO SEMANAL
# ============================================================================

dag_report = DAG(
    'apex_motors_weekly_report',
    default_args=default_args,
    description='Geração de relatório semanal',
    schedule_interval='0 8 * * 1',  # Executa toda segunda-feira às 8 AM
    catchup=False,
    tags=['apex-motors', 'reporting'],
)

def generate_weekly_report():
    """Gera relatório semanal de vendas"""
    print("Gerando relatório semanal...")
    # Implementar lógica de relatório
    print("Relatório gerado!")

report_task = PythonOperator(
    task_id='generate_weekly_report',
    python_callable=generate_weekly_report,
    dag=dag_report,
)
