requirements.txt
# Apache Airflow
apache-airflow==2.8.0
apache-airflow-providers-google==10.10.0
apache-airflow-providers-amazon==8.0.0

# Google Cloud
google-cloud-bigquery==3.14.0
google-cloud-bigquery-storage==2.22.0
google-cloud-storage==2.10.0

# Data Processing
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.21

# dbt
dbt-core==1.7.0
dbt-bigquery==1.7.0

# Utilities
python-dotenv==1.0.0
pydantic==2.0.3
requests==2.31.0

# Testing & Validation
pytest==7.4.2
great-expectations==0.17.12

# Logging & Monitoring
python-logging-loki==0.3.2
